tpm: Move functions from tpm_astract.py to tpm_util.py
Move functions from TPM_Utilities class from tpm_abstract.py into
tpm_util.py and drop the TPM_Utilities class.
